First and second stage

+++ src/main/kotlin/day11/GalaxyBrowse.kt
@@ -1,7 +1,7 @@
 package day11
 
 class GalaxyBrowse(private val galaxy:Galaxy) {
-    fun calculateSumOfShortestPath():UInt{
+    fun calculateSumOfShortestPath():ULong{
         return galaxy
             .expand()
             .assignName()
